refactor: move message service into folder
