+++ src/app/auth/components/login-form/login-form.component.spec.ts
@@ -7,7 +7,7 @@ import { of, throwError } from 'rxjs';
 
 import { LoginFormComponent } from '@/app/auth/components/login-form/login-form.component';
 import { AuthService } from '@/app/auth/services/auth/auth.service';
-import { MessageService as UserMessageService } from '@/app/shared/services/message.service';
+import { MessageService as UserMessageService } from '@/app/shared/services/message/message.service';
 
 describe('LoginFormComponent', () => {
   let component: LoginFormComponent;
